Add files via upload
The data Is not shared due to copyright from other author.

This code reads a csv file with data on the physical photoionization characteristics of a grid of planetary nebulae. In the code there is a function that receives the chosen parameters (such as element abundance or emission line fluxes) and plots a curve graph that generally uses the effective temperature of the central star as the x-axis and with this we create a temperature range for different characteristics of these bodies.